ecommerce_backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
seaborn==0.13.2
openpyxl==3.1.2
fpdf2==2.7.9
python-multipart==0.0.9
